refactor: core 패키지 game으로 rename 및 Applicatin 진입점 루트 디렉터리로 이동
